Documentation: using NAs in the RHS
Following [this issue](https://github.com/tidyverse/dplyr/issues/3202) I added a couple of sentences in the @param section intended to be helpful to new users, addressing directly a mistake that seems to be frequent. It's very rare for the common user to be aware of the different types of NA that exist in R (I use R since 2005 and today's the first time I've come across this).